+++ lib/transports/websocket/index.js
@@ -4,8 +4,8 @@
  */
 
 module.exports = {
-  7: require('./hybi-07-12'),
-  8: require('./hybi-07-12'),
-  13: require('./hybi-16'),
+  7: require('./hybi'),
+  8: require('./hybi'),
+  13: require('./hybi'),
   default: require('./default')
 };
