+++ package.json
@@ -19,6 +19,7 @@
         "socket.io-client": "0.8.7"
       , "policyfile": "0.0.4"
       , "redis": "0.6.7"
+      , "ws": "0.4.1"
     }
   , "devDependencies": {
         "expresso": "0.9.2"
